extract data and status socket to separate module
